+++ src/translate.js
@@ -1,12 +1,16 @@
 import { utf8ToBase64 } from './common/base64';
 
+// Current locales
 let _locales = {};
+
+// Caches
+const allLocales = {};
 const keyCache = {};
 
-const translate = (locales, utf8, data) => {
+const translate = (utf8, data) => {
   let base64 = keyCache[utf8];
   if (base64 == null) base64 = keyCache[utf8] = utf8ToBase64(utf8);
-  const fnTranslate = locales[base64];
+  const fnTranslate = _locales[base64];
   let out;
   if (fnTranslate) {
     out = fnTranslate(data);
@@ -21,13 +25,33 @@ const translate = (locales, utf8, data) => {
   return out;
 };
 
-const _t = (key, data) => translate(_locales, key, data);
+const _t = translate;
 
-_t.setLocales = locales => { _locales = locales; };
+_t.setLocales = langOrLocales => {
+  if (typeof langOrLocales === 'string') {
+    let lang = langOrLocales;
+    let locales = null;
+    while (!(locales = allLocales[lang])) {
+      lang = _t.getParentBcp47(lang);
+      if (!lang) return null;
+    }
+    _locales = locales;
+    return lang;
+  } else {
+    _locales = langOrLocales;
+    return null
+  }
+};
+
+_t.addLocales = (lang, locales) => { allLocales[lang] = locales; };
+
 _t.getParentBcp47 = bcp47 => {
   const tokens = bcp47.replace(/_/g, '-').split('-');
   if (tokens.length <= 1) return null;
   return tokens.slice(0, tokens.length - 1).join('-');
 };
 
+// =======================================
+// Public API
+// =======================================
 module.exports = _t;