+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+* [m] Add `_t.addLocales(lang, locales)` function and `_t.setLocales(lang)` signature: Mady can now keep track of all locales and use parent BCP47 codes when the requested BCP47 is not available.
+
 ## 1.5.0 (August 1, 2016)
 
 * [M] Better **BCP47** support: complete missing translations with those from descendants and ancestors.